handle requestAnimationFrame calls being missed
When the tab is in the background the browser stops calling
requestAnimation Frame. We now handle adding items that should would
have been added whilst the animation was suspended. If all current
items would have moved off screen whilst the animation was suspended
then we start the next item offscreen, otherwise we could end up with a
huge amount of items to discard and it’s pretty inefficient